autonomous_llm_agent/requirements.txt
python-dotenv
requests
openai>=1.0.0
httpx
